rackspace-email-management#13 Fix spam sync
The Spam "settings" sync will fail if the spam configuration
block is missing any subsection (`settings, `blocklist`, `ipblocklist`,
`safelist` or `ipsafelist`).  Allow a section to be missing
in configuration, indicating it is not managed.

This is not a complete solution, but addresses the immediate
failure.  Need more robust logic, will be an enhancement feature
later

---

Changes:
---------

`rackspace/spam.py`
* Skip a spam subsection if it is not present in the config

`sync.py`
* Minor changes to provide a wait loop for changes.  Meant for
  containerization, but not complete

---

Testing:
---------

Rerun the `sync.py` with a a partial spam account settings

---

Branch: `rackspace-email-management#13`